+++ applications/vessel-history/src/features/profile/components/activity/Activity.tsx
@@ -1,13 +1,14 @@
-import React, { Fragment, useCallback, useEffect, useState, useMemo } from 'react'
+import React, { Fragment, useCallback, useEffect, useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
+import AutoSizer from 'react-virtualized-auto-sizer'
+import { VariableSizeList as List } from 'react-window'
 import { Modal, Spinner } from '@globalfishingwatch/ui-components'
 import { selectResourcesLoading } from 'features/resources/resources.slice'
 import { VesselWithHistory } from 'types'
-import {
-  RenderedEvent,
-  selectEventsWithRenderingInfo,
-} from 'features/vessels/activity/vessels-activity.slice'
+import { RenderedEvent } from 'features/vessels/activity/vessels-activity.slice'
+import { selectFilteredEvents } from 'features/vessels/activity/vessels-activity.selectors'
 import { fetchRegionsThunk } from 'features/regions/regions.slice'
+import ActivityFilters from 'features/profile/filters/ActivityFilters'
 import ActivityItem from './ActivityItem'
 import ActivityModalContent from './ActivityModalContent'
 import styles from './Activity.module.css'
@@ -21,11 +22,7 @@ const Activity: React.FC<ActivityProps> = (props): React.ReactElement => {
   const dispatch = useDispatch()
 
   const eventsLoading = useSelector(selectResourcesLoading)
-  const eventsForTracks = useSelector(selectEventsWithRenderingInfo)
-  const events = useMemo(
-    () => eventsForTracks.flat().sort((a, b) => (a.start > b.start ? -1 : 1)),
-    [eventsForTracks]
-  )
+  const events = useSelector(selectFilteredEvents)
 
   const [isModalOpen, setIsOpen] = useState(false)
   const [selectedEvent, setSelectedEvent] = useState<RenderedEvent>()
@@ -41,6 +38,7 @@ const Activity: React.FC<ActivityProps> = (props): React.ReactElement => {
 
   return (
     <Fragment>
+      <ActivityFilters></ActivityFilters>
       {eventsLoading && <Spinner className={styles.spinnerFull} />}
       {!eventsLoading && (
         <Fragment>
@@ -48,15 +46,32 @@ const Activity: React.FC<ActivityProps> = (props): React.ReactElement => {
             {selectedEvent && <ActivityModalContent event={selectedEvent}></ActivityModalContent>}
           </Modal>
           <div className={styles.activityContainer}>
-            {/* TODO: Implement virtual rendering|filtering to boost performance and usability  */}
-            {events &&
-              events.map((event, eventIndex) => (
-                <ActivityItem
-                  key={eventIndex}
-                  event={event}
-                  onInfoClick={(event) => openModal(event)}
-                />
-              ))}
+            {events && events.length > 0 && (
+              <AutoSizer disableWidth={true}>
+                {({ width, height }) => (
+                  <List
+                    width={width}
+                    height={height}
+                    itemCount={events.length}
+                    itemData={events}
+                    itemSize={() => 79}
+                  >
+                    {({ index, style }) => {
+                      const event = events[index]
+                      return (
+                        <div style={style}>
+                          <ActivityItem
+                            key={index}
+                            event={event}
+                            onInfoClick={(event) => openModal(event)}
+                          />
+                        </div>
+                      )
+                    }}
+                  </List>
+                )}
+              </AutoSizer>
+            )}
           </div>
         </Fragment>
       )}