+++ applications/vessel-history/src/features/profile/components/activity/ActivityModalContentDetailsEncounter.tsx
@@ -1,10 +1,14 @@
 import React, { Fragment } from 'react'
 import { useTranslation } from 'react-i18next'
 import { EventVessel } from '@globalfishingwatch/api-types/dist'
-import { ActivityEvent } from 'types/activity'
+import { DEFAULT_EMPTY_VALUE } from 'data/config'
+import { RenderedEvent } from 'features/vessels/activity/vessels-activity.slice'
+import ActivityModalContentField from './ActivityModalContentField'
+import ActivityModalContentDetails from './ActivityModalContentDetails'
+import styles from './ActivityModalDetails.module.css'
 
 interface ActivityModalContentProps {
-  event: ActivityEvent
+  event: RenderedEvent
 }
 
 const ActivityModalContentDetailsEncounter: React.FC<ActivityModalContentProps> = (
@@ -16,20 +20,40 @@ const ActivityModalContentDetailsEncounter: React.FC<ActivityModalContentProps> 
 
   return (
     <Fragment>
-      {event.encounter && <p>
-        <span>{t('event.medianSpeed', 'Median Speed')}:</span>
-        {event.encounter.medianSpeedKnots.toFixed(2)}knots
-        </p>
-      }
-      {event.encounter && <p><span>{t('event.authStatus', 'Authorization status')}:</span> {event.encounter.authorizationStatus}</p>}
-      {relatedVessel &&
-        <div>
-          <h3>{t('event.vesselInvolved', 'Vessel involved in the event')}:</h3>
-          <p><span>{t('vessel.name', 'Name')}:</span> {relatedVessel.name}</p>
-          <p><span>{t('vessel.flag', 'Flag')}:</span> {relatedVessel.flag}</p>
-          <p><span>{t('event.nextPort', 'Next port traveled')}:</span> {relatedVessel.nextPort?.label}</p>
+      {relatedVessel && (
+        <div className={styles.row}>
+          <ActivityModalContentField
+            label={t('vessel.encounteredVessel', 'Encountered Vessel')}
+            value={relatedVessel.name}
+          />
+          <ActivityModalContentField label={t('vessel.flag', 'Flag')} value={relatedVessel.flag} />
+          <ActivityModalContentField
+            label={t('event.nextPort', 'Next port traveled')}
+            value={relatedVessel.nextPort?.label}
+          />
         </div>
-      }
+      )}
+
+      <ActivityModalContentDetails event={event} />
+
+      {event.encounter && (
+        <ActivityModalContentField
+          label={t('event.medianSpeed', 'Median Speed')}
+          value={
+            event.encounter.medianSpeedKnots
+              ? t('event.formatSpeedKnots', '{{value}} knots', {
+                  value: event.encounter.medianSpeedKnots.toFixed(2),
+                })
+              : DEFAULT_EMPTY_VALUE
+          }
+        />
+      )}
+      {event.encounter && (
+        <ActivityModalContentField
+          label={t('event.authStatus', 'Authorization status')}
+          value={event.encounter.authorizationStatus}
+        />
+      )}
     </Fragment>
   )
 }