+++ applications/vessel-history/src/features/profile/components/activity/ActivityModalContentDetailsFishing.tsx
@@ -1,10 +1,13 @@
 import React, { Fragment } from 'react'
 import { useTranslation } from 'react-i18next'
 import { DEFAULT_EMPTY_VALUE } from 'data/config'
-import { ActivityEvent } from 'types/activity'
+import { RenderedEvent } from 'features/vessels/activity/vessels-activity.slice'
+import ActivityModalContentDetails from './ActivityModalContentDetails'
+import ActivityModalContentField from './ActivityModalContentField'
+import styles from './ActivityModalDetails.module.css'
 
 interface ActivityModalContentProps {
-  event: ActivityEvent
+  event: RenderedEvent
 }
 
 const ActivityModalContentDetailsFishing: React.FC<ActivityModalContentProps> = (
@@ -15,19 +18,29 @@ const ActivityModalContentDetailsFishing: React.FC<ActivityModalContentProps> = 
 
   return (
     <Fragment>
-      <p>
-        <span>{t('event.distance', 'Distance')}:</span>{' '}
-        {event.fishing?.totalDistanceKm
-          ? t('event.formatDistanceKm', '{{value}} km', {
-              value: event.fishing?.totalDistanceKm.toFixed(2),
-            })
-          : DEFAULT_EMPTY_VALUE}
-      </p>
-      <p>
-        <span>{t('event.avgSpeed', 'Avg Speed')}:</span>{' '}
-        {event.fishing?.averageSpeedKnots.toFixed(2)}
-        {t('event.knots', 'knots')}
-      </p>
+      <div className={styles.row}>
+        <ActivityModalContentField
+          label={t('event.distance', 'Distance')}
+          value={
+            event.fishing?.totalDistanceKm
+              ? t('event.formatDistanceKm', '{{value}} km', {
+                  value: event.fishing?.totalDistanceKm.toFixed(2),
+                })
+              : DEFAULT_EMPTY_VALUE
+          }
+        />
+        <ActivityModalContentField
+          label={t('event.avgSpeed', 'Avg Speed')}
+          value={
+            event.fishing?.averageSpeedKnots
+              ? t('event.formatSpeedKnots', '{{value}} knots', {
+                  value: event.fishing?.averageSpeedKnots.toFixed(2),
+                })
+              : DEFAULT_EMPTY_VALUE
+          }
+        />
+      </div>
+      <ActivityModalContentDetails event={event} />
     </Fragment>
   )
 }