+++ applications/vessel-history/src/types/index.ts
@@ -34,6 +34,7 @@ export interface MapCoordinates extends CoordinatePosition {
 export interface FieldValueCounter<T = string> {
   counter: number
   name: T
+  source?: VesselAPISource
 }
 
 export interface VesselFieldHistory<T> {
@@ -68,6 +69,7 @@ export interface ValueItem<T = string> {
   value: T
   firstSeen?: string
   endDate?: string
+  source?: VesselAPISource
 }
 export type BuiltYear = ValueItem
 export type Flag = ValueItem